backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python-headless==4.9.0.80
ultralytics==8.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.3
fpdf2==2.7.7
cryptography

